src/services/payment/requirements.txt
grpcio
grpcio-tools
asyncpg
pydantic
motor